Workbook Overhaul
I completely overhauled the Auditing Sheet. I wanted to make it faster and more user-friendly. First, I made a "Create Auditing Sheet" workbook. In this workbook, the user will find a button called "Create New Auditing Sheet." When clicked, a user-form will appear; and the user can easily make a new worksheet for the day. Everything is automatic. Secondly, I overhauled the code of the macros, making the workbook much faster than it used to be.