Add files via upload
This will start a new line of Auditing Sheets. I completely nuked the VBA code, and started over. I also added in a new worksheet that shows individual errors. This is still in designing mode. And, as I've mentioned before, this auditing sheet is not meant to be used anywhere else. It is mainly for learning purposes, and will hopefully give you ideas on how to create your own.